스켈레톤 적용, progress 컴포넌트 분리
